Refactor location tracking and update Android version
Updated `BackgroundLocationTracking.csproj` to change the minimum supported Android version from 21.0 to 26.0.

In `MainPage.xaml`, removed Esri MapView and ViewModel binding, added buttons for starting/stopping location tracking, and updated log message Editor.

In `MainPage.xaml.cs`, added Esri location services references, introduced `LocationDataSource`, and added methods for location tracking with permission checks and event handling.
